chore: 특정 모델 변경 팝업 dir 이동
